Add the initial structure for the server
The change introduces the initial structure of the server providing the basic functionality. Namely, it includes the logic for establishing the server, handling page routing, database interactions, security configurations, websockets, and HTTP. Also added a ktor library for server instantiation. Along with that, some utilities and gradle files for the build process have been incorporated.