Introduced chat feature in the game
Added ChatUsecase, ChatRepository and related classes to facilitate communication between players in a game room. This feature allows users to send messages to all opponents in the same room. If the user is not joined to any room and tries to send a message, they will be notified. The purpose of this addition is to improve in-game communication for a better user experience. The refactor also moves connection maintenance to BattleDependencies for consistency.