Added a PluginRegistry to handle application plugins
A PluginRegistry was introduced in order to manage plugins in application layers. Each plugin was represented as a class extending from an abstract class 'Plugin.' A new method 'install' adds a plugin to a mutable list of plugins in the PluginRegistry. This was further integrated into the frontend where each incoming frame is now processed by the installed plugins via the 'onMessage' method of each plugin, and if the message satisfies certain condition, a certain action is taken. For example, a 'QuitPlugin' was added which closes the session when a quitting command is received. This change was done to modularize and make the handling of different features and commands more cohesive and maintainable.