Refactor code and add room leaving functionality.
This commit includes several changes, all seeking to improve the clarity and capability of the project. The `onMessage` method was simplified to use a lambda for better readability. Also, this function's scope was clarified by changing it from `var` to `internal var`.

Moreover, the mechanism to allow leaving a room was introduced. A new file `LeaveRoomRepository.kt` was created to handle the logic of a user leaving a room, and its usage was integrated into `Routing.kt`. The frontend part was updated accordingly, removing the todo tag related to the room leaving logic.

These changes enhance the project's structure and widen the functionality range, taking a step towards a more complete application.